Move Facetline out of header
Because it doesn't need to be part of the header and to avoid irrelevant prop drilling.  I have renamed the facetLine component to facetLineFilters to be able to create a new container component called facetLine
