+++ src/apps/search-result/search-result.tsx
@@ -26,6 +26,7 @@ import {
 } from "../../core/dpl-cms/model";
 import Campaign from "../../components/campaign/Campaign";
 import FacetBrowserModal from "../../components/facet-browser/FacetBrowserModal";
+import FacetLine from "../../components/facet-line/FacetLine";
 
 interface SearchResultProps {
   q: string;
@@ -122,12 +123,8 @@ const SearchResult: React.FC<SearchResultProps> = ({ q, pageSize }) => {
 
   return (
     <div className="search-result-page">
-      <SearchResultHeader
-        hitcount={String(hitcount)}
-        q={q}
-        filters={filters}
-        filterHandler={filteringHandler}
-      />
+      <SearchResultHeader hitcount={String(hitcount)} q={q} />
+      <FacetLine q={q} filters={filters} filterHandler={filteringHandler} />
       {campaignData && campaignData.data && (
         <Campaign campaignData={campaignData.data} />
       )}
