+++ src/components/facet-line/FacetLine.tsx
@@ -1,120 +1,43 @@
-import { upperFirst } from "lodash";
 import React from "react";
-import { useDispatch } from "react-redux";
-
 import {
   FilterItemTerm,
   TermOnClickHandler
 } from "../../apps/search-result/types";
 import {
   FacetResult,
-  FacetValue
+  useIntelligentFacetsQuery
 } from "../../core/dbc-gateway/generated/graphql";
-import { openModal } from "../../core/modal.slice";
-import { useText } from "../../core/utils/text";
-import ButtonTag from "../Buttons/ButtonTag";
-import { FacetBrowserModalId } from "../facet-browser/FacetBrowserModal";
-import Dropdown from "./Dropdown";
+import FacetLineSelected from "./FacetLineSelected";
+import FacetLineFilters from "./FacetLineFilters";
 
 type FacetLineProps = {
-  facets: FacetResult[];
+  q: string;
   filters: { [key: string]: { [key: string]: FilterItemTerm } };
   filterHandler: TermOnClickHandler;
 };
 
 const FacetLine: React.FunctionComponent<FacetLineProps> = ({
-  facets,
+  q,
   filters,
   filterHandler
 }) => {
-  const t = useText();
-  const dispatch = useDispatch();
-
-  const formatValuesToDropdown = (values: FacetValue[]) => {
-    return values.map((value) => {
-      return {
-        title: value.term,
-        value,
-        disabled: Boolean(filters?.genreAndForm?.[value.term]),
-        selected: Boolean(filters?.genreAndForm?.[value.term])
-      };
-    });
-  };
-
-  const handleDropdownOnchange = (
-    e: React.ChangeEvent<HTMLSelectElement>,
-    facet: string
-  ) => {
-    const term = facets
-      .find((item) => item.name === facet)
-      ?.values.find((item) => item.term === e.target.value) as FilterItemTerm;
-
-    filterHandler({
-      filterItem: {
-        facet,
-        term
-      },
-      action: "add"
-    });
-  };
+  const { data } = useIntelligentFacetsQuery({
+    q: { all: q },
+    facetsLimit: 5,
+    valuesLimit: 5
+  });
 
   return (
-    <ul className="facet-line mt-48">
-      {facets.map(({ name, values }) => {
-        if (values.length > 1) {
-          return (
-            <li className="facet-line__item">
-              <Dropdown
-                placeholder={t(`facet${upperFirst(name)}Text`)}
-                list={formatValuesToDropdown(values)}
-                ariaLabel={name}
-                arrowIcon="chevron"
-                classNames="dropdown--grey-borders"
-                innerClassNames={{
-                  select: "dropdown__select--inline",
-                  arrowWrapper: "dropdown__arrows--inline "
-                }}
-                handleOnChange={(e) => handleDropdownOnchange(e, name)}
-              />
-            </li>
-          );
-        }
-
-        return (
-          <>
-            {values.map((termObj) => {
-              const { term, score } = termObj;
-              return (
-                <li className="facet-line__item">
-                  <ButtonTag
-                    key={term}
-                    onClick={() =>
-                      filterHandler({
-                        filterItem: {
-                          facet: name,
-                          term: termObj as FilterItemTerm
-                        },
-                        action: "add"
-                      })
-                    }
-                    removeAriaPressed={Boolean(filters?.[name]?.[term])}
-                  >
-                    {`${term} (${score})`}
-                  </ButtonTag>
-                </li>
-              );
-            })}
-          </>
-        );
-      })}
-      <li className="facet-line__item">
-        <ButtonTag
-          onClick={() => dispatch(openModal({ modalId: FacetBrowserModalId }))}
-        >
-          {t("addMoreFiltersText")}
-        </ButtonTag>
-      </li>
-    </ul>
+    <>
+      {data && (
+        <FacetLineFilters
+          filters={filters}
+          facets={data.search.intelligentFacets as FacetResult[]}
+          filterHandler={filterHandler}
+        />
+      )}
+      <FacetLineSelected filters={filters} filterHandler={filterHandler} />
+    </>
   );
 };
